fix(order): 修复JWT用户ID类型转换和删除购物车接口参数问题
1. 修复JWT中间件用户ID类型：float64 -> int，保持类型一致性
2. 统一订单模块所有接口使用middlewares.KeyUserID常量
3. 修复删除购物车接口参数传递：通过购物车ID查找商品ID后删除
4. 解决DeleteCartItem gRPC调用参数不匹配问题

+++ internal/app/emshop/api/auth.go
@@ -36,7 +36,7 @@ func (c *customJWTAuth) AuthFunc() gin.HandlerFunc {
 			return
 		}
 
-		ctx.Set(middlewares.KeyUserID, float64(claims.ID))
+		ctx.Set(middlewares.KeyUserID, int(claims.ID))
 		ctx.Next()
 	}
 }
